+++ Github/Repos/Collaborators.hs
@@ -21,18 +21,18 @@ import Github.Request
 -- | All the users who have collaborated on a repo.
 --
 -- > collaboratorsOn "thoughtbot" "paperclip"
-collaboratorsOn :: Name GithubOwner -> Name Repo -> IO (Either Error (Vector SimpleOwner))
+collaboratorsOn :: Name GithubOwner -> Name Repo -> IO (Either Error (Vector SimpleUser))
 collaboratorsOn = collaboratorsOn' Nothing
 
 -- | All the users who have collaborated on a repo.
 -- With authentication.
-collaboratorsOn' :: Maybe GithubAuth -> Name GithubOwner -> Name Repo -> IO (Either Error (Vector SimpleOwner))
+collaboratorsOn' :: Maybe GithubAuth -> Name GithubOwner -> Name Repo -> IO (Either Error (Vector SimpleUser))
 collaboratorsOn' auth user repo =
     executeRequestMaybe auth $ collaboratorsOnR user repo Nothing
 
 -- | List collaborators.
 -- See <https://developer.github.com/v3/repos/collaborators/#list-collaborators>
-collaboratorsOnR :: Name GithubOwner -> Name Repo -> Maybe Count -> GithubRequest k (Vector SimpleOwner)
+collaboratorsOnR :: Name GithubOwner -> Name Repo -> Maybe Count -> GithubRequest k (Vector SimpleUser)
 collaboratorsOnR user repo =
     GithubPagedGet ["repos", toPathPart user, toPathPart repo, "collaborators"] []
 
@@ -44,7 +44,7 @@ collaboratorsOnR user repo =
 isCollaboratorOn :: Maybe GithubAuth
                  -> Name GithubOwner  -- ^ Repository owner
                  -> Name Repo         -- ^ Repository name
-                 -> Name GithubOwner  -- ^ Collaborator?
+                 -> Name User         -- ^ Collaborator?
                  -> IO (Either Error Bool)
 isCollaboratorOn auth user repo coll =
     executeRequestMaybe auth $ isCollaboratorOnR user repo coll
@@ -53,7 +53,7 @@ isCollaboratorOn auth user repo coll =
 -- See <https://developer.github.com/v3/repos/collaborators/#check-if-a-user-is-a-collaborator>
 isCollaboratorOnR :: Name GithubOwner  -- ^ Repository owner
                   -> Name Repo         -- ^ Repository name
-                  -> Name GithubOwner  -- ^ Collaborator?
+                  -> Name User         -- ^ Collaborator?
                   -> GithubRequest k Bool
 isCollaboratorOnR user repo coll = GithubStatus StatusOnlyOk $
     GithubGet ["repos", toPathPart user, toPathPart repo, "collaborators", toPathPart coll] []